pyproject.toml
[project]
name = "realtime-fraud-detection-mlops"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.64.2",
    "fastapi>=0.121.2",
    "feast>=0.57.0",
    "scikit-learn>=1.5.0",
    "imbalanced-learn>=0.12.3",
    "mlflow>=3.6.0",
    "prefect==3.6.9",
    "pytest>=9.0.1",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.2",
    "shap>=0.46.0",
    "uvicorn>=0.34.0",
    "xgboost==2.0.3",
    "great-expectations[cli]>=1.10.0",
    "evidently==0.7.18",
    "pydantic-settings>=2.12.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "boto3<1.42.20",
    "s3fs>=0.4.2",
    "python-dotenv>=1.2.1",
    "prefect-kubernetes>=0.7.4",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
]

[tool.ruff]
# Set line length to 100 (not default 88)
line-length = 100
target-version = "py312"

# Exclude directories
exclude = [
    ".venv",
    "__pycache__",
    "*.pyc",
    "gx",
    "notebooks",
    "docker",
    "k8s",
    ".pytest_cache",
    "dist",
    "build",
    "mlruns",
]

[tool.ruff.lint]
# Enable these rule groups
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "UP",  # pyupgrade (modern Python syntax)
]

# Ignore specific rules that conflict with MLflow/Prefect patterns
ignore = [
    "E501",  # Line too long (we handle with line-length setting)
]

[tool.ruff.lint.isort]
# Import sorting configuration
known-first-party = ["src"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]